split-out default.rb into recipes and helpers module
