+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'irving@chef.io'
 license 'all_rights'
 description 'Installs/Configures postgresql-cluster'
 long_description 'Installs/Configures postgresql-cluster'
-version '0.2.0'
+version '0.3.0'
 source_url       'https://github.com/irvingpop/postgresql-cluster'
 issues_url       'https://github.com/irvingpop/postgresql-cluster/issues'
